Hoist formData state to App level
