+++ frontend/src/App.js
@@ -5,12 +5,16 @@ import MyTripsPage from "./MyTripsPage"
 import CreateNewTripPage from "./CreateNewTripPage"
 import "bootstrap-icons/font/bootstrap-icons.css"
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom"
-
+const initialValue = {
+  title: null,
+  start_date: null,
+  end_date: null,
+  location_id: null,
+}
 function App() {
   const [trips, setTrips] = useState([])
   const [locations, setLocations] = useState([])
-  // const [activities, setActivities] = useState([])
-  // const [lodgings, setLodgings] = useState([])
+  const [formData, setFormData] = useState(initialValue)
 
   function onAddNewTrip(trip) {
     setTrips([...trips, trip])
@@ -28,18 +32,6 @@ function App() {
       .then((locations) => setLocations(locations))
   }, [])
 
-  // useEffect(() => {
-  //   fetch("http://localhost:9292/activities")
-  //     .then((r) => r.json())
-  //     .then((activities) => setActivities(activities))
-  // }, [])
-
-  // useEffect(() => {
-  //   fetch("http://localhost:9292/lodgings")
-  //     .then((r) => r.json())
-  //     .then((lodgings) => setLodgings(lodgings))
-  // }, [])
-
   return (
     <Router>
       <div className="App">
@@ -52,6 +44,8 @@ function App() {
                 <CreateNewTripPage
                   locations={locations}
                   onAddNewTrip={onAddNewTrip}
+                  formData={formData}
+                  setFormData={setFormData}
                 />
               }
             />
